Moved FunBase & FilterBase to Models
